src/main/java/com/brightsparklabs/asanti/model/schema/typedefinition/AsnSchemaTypeDefinitionGeneralizedTime.java
/*
 * Created by brightSPARK Labs
 * www.brightsparklabs.com
 */

package com.brightsparklabs.asanti.model.schema.typedefinition;

import com.brightsparklabs.asanti.model.schema.AsnBuiltinType;
import com.brightsparklabs.asanti.model.schema.constraint.AsnSchemaConstraint;

/**
 * A {@code GeneralizedTime} type definition from a within a module specification within an ASN.1
 * schema.
 *
 * @author brightSPARK Labs
 */
public class AsnSchemaTypeDefinitionGeneralizedTime extends AsnSchemaTypeDefinitionPrimitive
{
    // -------------------------------------------------------------------------
    // CONSTRUCTION
    // -------------------------------------------------------------------------

    /**
     * Default constructor.
     *
     * @param name
     *         name of the GeneralizedTime type definition
     * @param constraint
     *         The constraint on the type. Use {@link AsnSchemaConstraint#NULL} if no constraint.
     *
     * @throws NullPointerException
     *         if {@code name} is {@code null}
     * @throws IllegalArgumentException
     *         if {@code name} is blank
     */
    public AsnSchemaTypeDefinitionGeneralizedTime(String name, AsnSchemaConstraint constraint)
    {
        super(name, AsnBuiltinType.GeneralizedTime, constraint);
    }

    // -------------------------------------------------------------------------
    // IMPLEMENTATION: AsnSchemaTypeDefinition
    // -------------------------------------------------------------------------

    @Override
    public Object visit(AsnSchemaTypeDefinitionVisitor<?> visitor)
    {
        return visitor.visit(this);
    }
}
